Updated website based on information from Marikia
Updated text, and created placeholders for rooms 1-10 for the Elsternwick location. The About page still only reflects the St. Kilda Location, images need to be implemented as well. Previous rooms descritpion mention number of chairs also, but there is no mention of that in the email. I am assuming each one will have the same number of chairs anyway.